[MOS-788] Fix incorrect total CPU usage in logs
After a long time, a counter overflow appeared

+++ module-sys/SystemManager/PowerManager.cpp
@@ -32,7 +32,7 @@ namespace sys
     [[nodiscard]] auto CpuFrequencyMonitor::GetTotalRuntimePercentage(
         const TickType_t totalTicksIncrease) const noexcept -> std::uint32_t
     {
-        return totalTicksIncrease == 0 ? 0 : ((totalTicksCount * 100) / totalTicksIncrease);
+        return totalTicksIncrease == 0 ? 0 : ((static_cast<std::uint64_t>(totalTicksCount) * 100) / totalTicksIncrease);
     }
 
     [[nodiscard]] auto CpuFrequencyMonitor::GetPeriodRuntimePercentage(
